AnimatedTextureRect: Default sprite_frames to null
Previously, this class defaulted to showing the template player, by
path.

Our convention is to preload by uid:// in scripts to avoid needing to
update every script when we rename a resource.

But more to the point: TextureRect does not display anything by default;
and this is meant to be an animated version of that standard Control.