splash, title: Reference next scene by path, not as PackedScene
Previously, splash.gd preload()ed (the UID path that resolves to)
res://scenes/menus/title/title_screen.tscn, and title_screen.tscn
referenced intro.tscn as a PackedScene.

When I wrote this code I did not fully appreciate the consequences of
using preload()/PackedScene in these kinds of contexts.

Doing this means that to load the splash screen scene, the title screen
scene must also be loaded (though not instantiated), which is not
necessary to show the splash screen. And to load the title screen scene,
it is necessary to load the intro scene.

As well as doing more up-front work than is strictly necessary, this
actually causes problems when you first open a fresh clone of the
project in the editor. In this case, the default scene is loaded: this
is splash.tscn. As described above, this transitively loads intro.tscn;
that scene uses cinematic.gd; and cinematic.gd preload()s
cinematic_placeholder.dialogue.

But at this point during the initial import, Godot Dialogue Manager has
not finished loading, and so the `importer="dialogue_manager"` line in
cinematic_placeholder.dialogue.import causes a cascade of errors:

    ERROR: BUG: File queued for import, but can't be imported, importer
           for type 'dialogue_manager' not found.

    ERROR: Failed loading resource:
           res://scenes/ui_elements/cinematic/cinematic_placeholder.dialogue.
           Make sure resources have been imported by opening the project
           in the editor at least once.

    ERROR: res://scenes/ui_elements/cinematic/cinematic.gd:9 - Parse
           Error: Could not preload resource file "uid://b7ad8nar1hmfs".

    ERROR: modules/gdscript/gdscript.cpp:3022 - Failed to load script
           "res://scenes/ui_elements/cinematic/cinematic.gd" with error
           "Parse error".

    ERROR: BUG: File queued for import, but can't be imported, importer
           for type 'dialogue_manager' not found.

    ERROR: Failed loading resource:
           res://scenes/menus/intro/components/intro.dialogue. Make sure
           resources have been imported by opening the project in the
           editor at least once.

The next time you open the project, everything works correctly because
some cached state about the imported resource persists in `.godot`.

If the default scene used a DialogueResource, I don't immediately see a
way to avoid this; but fortunately it does not.

So: replace the PackedScene constants in splash.gd and title_screen.gd
with Strings. This follows the same pattern as we use in cinematic.gd
and collectible_item.gd to reference the next scene (which was necessary
to avoid a cycle); reduces the amount of work that must be done up-front
to load and display the splash and title screens; and prevents the
cascade of errors when loading splash.tscn on first launch of the
editor.

While doing this, don't define the values in the `.gd` file: make them exported
properties and set them in the corresponding `.tscn`. I think this is the best
of both worlds because the real path is then shown in the inspector.

Fixes https://github.com/endlessm/threadbare/issues/1216