Add development status to quest resource (#1155)
This adds a new field to the Quest resource, indicating whether the
quest is in development, fully implemented, or broken.

This information will be used in future to label quests in the
storybook.

All quests currently in the tree are either in development, or complete.
In particular, the template `quest.tres` is labelled "in development"
because this file is duplicated for new storyquests.

Resolves #750

Co-authored-by: Will Thompson <wjt@endlessaccess.org>