pre-commit: Exclude all image files from trailing whitespace checks
Previously we excluded SVG files from these checks with a regex matching
SVG filenames.

I learnt that pre-commit uses the “identify” library to determine a set
of types for a file. This library already knows that `.svg` files are
images:

    $ identify-cli addons/git_describe/label/icon.svg
    ["file", "image", "non-executable", "svg", "text", "xml"]

SVG is not the only text-based image format. We don't want to run
whitespace checks on any kind of image file.

Exclude image files by type, not SVGs by filename.

https://pypi.org/project/identify/