pre-commit: Exclude Piskel files from whitespace checks
Like SVGs, these are stored as plain text, but are not meant to be hand-edited.

The “identify” library doesn't recognise *.piskel files as images. Match
them by filename.

I submitted https://github.com/pre-commit/identify/pull/541 to teach
identify about these files; if that change is accepted and released this
could be reverted.