pre-commit: Disable style checks for StoryQuests
We enforce coding style and formatting checks on the project as a whole
because we want the reusable code in this project to be exemplary, and
we believe that consistent style and formatting helps to maintain code
quality – and to avoid discussions about style and formatting!

There is less need to apply the same checks to scripts, shaders, and dialogue
files in StoryQuests: these are only used within that specific StoryQuest,
rather than being intended for other quests or components to reuse or extend.
The exceptions is the template quest, and Stella: we do want to maintain
consistent style there.

Exclude all storyquests, except the template and stella, from GDScript style
checks and editorconfig formatting checks. Keep other checks enabled, such as
for large files that should be stored in Git LFS.

This is currently mostly hypothetical – none of the StoryQuests in the tree have
any scripts or shaders, just dialogue files – but there are other quests being
developed which do include custom scripts and shaders.

Fixes https://github.com/endlessm/threadbare/issues/1227