Auto-detect SequencePuzzle steps (#1196)
Previously, the SequencePuzzle's `steps` property had to be set in the
editor to a list of `SequencePuzzleStep` nodes. This turned out to be
error-prone: it was easy to add a new `SequencePuzzleStep` node but
forget to add it to the `steps` property.

Instead, use a similar approach to how `SequencePuzzleObject` nodes are
discovered: find all `SequencePuzzleStep` nodes that are children (or
grandchildren, etc.) of the `SequencePuzzle` node. For
backwards-compatibility, only do this if the `steps` property is empty.

Resolves: https://github.com/endlessm/threadbare/issues/696

---------

Co-authored-by: Will Thompson <wjt@endlessaccess.org>