Void runner: Add idle motion to enemy
Previously the enemy/antagonist/void-spreading entity was invisible until
triggered. This made for a nice jump-scare but it seems a bit unfair that there
was no warning about what would happen if you crossed an invisible threshold.

Instead, introduce an idle state where the enemy follows a path. Place the path
just off the lower coast of the starting island, so it is visible at the start
of the level. This adds an ominous feel (I hope).

You might think that this would make the level too easy -- just wait until the
enemy is at the farthest-away point of its patrol before starting to run -- but
actually it doesn't because the enemy can cross the void. That's actually why I
designed the course to double back on itself -- I wanted the player to have to
think about their own (more constrained) route versus the pursuer's
straight-line route -- but it rarely happened. But now it does: if you try to
get a head-start, it just flies across the void to catch you up.

However, this adds a new possibility: the void may be crossing an existing
region of void, in which case it's invisible. To handle this, emit particles
after a certain distance (the diagonal size of a tile) has been travelled since
the last emission. This takes effect both when the enemy is catching up with
you, and also when it has walked its idle path and has doubled back on itself.

Initially there are tiles for the void to consume, but after it has walked the
path once there are not. Previously it would be invisible in this state.

https://github.com/endlessm/threadbare/issues/1237