CharacterSpriteBehavior: Automatically assign character
In the same way as we assign sprite if not set and the parent is an
AnimatedSprite2D, do it to the parent's parent if it is a
CharacterBody2D.

The rationale is that usually this is what you'll have in the Scene
Tree, while at the same time it keeps being configurable.

Also update the format in the existing documentation comment.