Default to fullscreen in non-web, non-editor builds
Previously, the default window mode was windowed. This brings with it some problems:

- On hidpi displays, the window is tiny;
- On small screens like Steam Deck, the window doesn't fill the screen
  but doesn't fit above the taskbar.

An easy fix for both, recommended in the documentation, is to fullscreen
by default. Make this change, using WINDOW_MODE_EXCLUSIVE_FULLSCREEN
(which as far as I can tell from the documentation is generally
preferable to WINDOW_MODE_FULLSCREEN).

However there are some complications:

- If a game is fullscreen by default, you cannot use the embedded mode
  in the editor.
- On the web, we can't currently fullscreen on startup: you can only
  fullscreen in response to a user gesture.

Rewrite the fullscreen setting to use project settings overrides, rather
than fullscreening on startup. This is the approach explained in
https://prvaak.cz/blog/window-mode-switching-in-godot/, a very helpful
post.

Then, use ProjectSettings' support for per-platform overrides to set the
default when running in the editor or on the web to windowed, as before.
These settings apparently can't be seen in the editor's Project Setting
dialog but will be preserved when the project is saved.

Resolves https://github.com/endlessm/threadbare/issues/1200