Gym: Use CharacterAnimationPlayerBehavior
Add it to the character that has an InputWalkBehavior. Also add an
AnimationPlayer node to it, with idle, walk and run animations. The walk
and run animations have rudimentary audio tracks. Add an
AudioStreamPlayer2D node for those sound effects.

Keep the CharacterSpriteBehavior but disable Play Animations in it, so
the character still flips horizontally.