Only mark quest as completed when Sokobans are solved
Previously, the quest was marked as complete once you took the threads to the
loom. But there's still the Sokoban puzzles to complete at that point:
conceptually the Sokobans are part of the quest.

Instead only mark the quest as complete once you have returned from the Sokobans
and the Elders have congratulated you.