GameState: Don't crash in mark_quest_completed() if not on a quest
I had previously misunderstood the ConfigFile.get_value() behaviour: in fact if
the third parameter (fallback value) is null (its default if not specified),
then if the key is missing it raises an error.

Instead pass empty string as default.

This case should never be hit because we only call mark_quest_completed() while
on a quest; but the function is designed to be safe to call when there is no
quest, so let's make it so.