Add an “Abandon Quest” option to the pause menu
This is only shown if the player is currently on a quest, including the
Sokoban puzzles. Update the button visibility each time the pause menu
is shown.

If chosen, the inventory is cleared, as is the incorporating_threads
flag that is set while inside the Eternal Loom; then Fray's End is
loaded. If Fray's End happens to already be loaded, that's OK: it will
be reloaded.

Resolves https://github.com/endlessm/threadbare/issues/432