Fray's End: Add cliff tiles below grass
Previously, there was no cliff below the bottom border of the top
island, so if you looked closely you could see the grey default
background colour peeking through.

The cliffs in the row of tiles below the edge of the grass were
previously part of Stone_Level1, which is above Grass_Level1.

Add a new, lower layer, Stone_Level0. Move the cliffs that are
conceptually below the grass from Stone_Level1 to Stone_Level0. Add
clifftop below the edge of the grass.

Resolves https://github.com/endlessm/threadbare/issues/1117