Fix ObjectDB & StringName leaks on exit
Previously, quitting the game from the main menu would show (under
`--verbose`):

    WARNING: ObjectDB instances leaked at exit (run with --verbose for details).
        at: cleanup (core/object/object.cpp:2378)
    Leaked instance: GDScriptFunctionState:9223372096078349844
    Hint: Leaked instances typically happen when nodes are removed from the scene tree (with `remove_child()`) but not freed (with `free()` or `queue_free()`).
    Orphan StringName: quit (static: 0, total: 1)
    Orphan StringName: finished (static: 0, total: 1)
    Orphan StringName: Music (static: 0, total: 1)
    Orphan StringName: _introduce_scene (static: 0, total: 1)
    StringName: 4 unclaimed string names at exit.

`quit` is the name of the method that quits the game; `finished` and
`_introduce_scene` are members of `transitions.gd` which are used at the
end of `do_transition`. Since commit 3176904 we quit midway through
`do_transition`.

Add a new method, `do_out_transition`, which only does the fade-out and
leaves the black cover layer visible. When quitting the game, await this
method, then call quit(), deferred, at the end of the frame.

This fixes the leaks above. (There are still an error about a leaked
texture, which appears to be unrelated.)

Resolves https://github.com/endlessm/threadbare/issues/1197