Add tomato, carrot, barley prop scenes
Previously, these were placed in Fray's End by instancing bush.tscn,
then setting its sprite_frames and scale.

Make a standalone scene for each prop, so that the sprite and scale can be set in one place.

Set the scale factor to be the same in both axes, and a power of 2: this
looks better and reduces jitter when moving the camera.

Replace all instances in Fray's End with these scenes. Group them under
new parent nodes.