Factor AnimatedSprite2D frame randomisation
For many props in the game, we want to randomise the start frame of the
animation, so that many nearby instances don't move in sync.

Previously we had 2 implementation of this: one in tree.gd and one in
animated_decoration.gd. Instead, add a new script which can be attached
to any AnimatedSprite2D to give it this behaviour. Update all relevant
scenes to use it.