Intro: Simplify waiting for transition to finish
We want the intro dialogue to not start until the fade-in from the previous
scene has finished.

Previously this was done by playing an animation that does nothing for 1 second.
We could instead use Godot Dialogue Manager's built-in syntax for waiting:

    do wait(1.0)

but we can do even better. Wait for the Transitions.finished signal, but ony if
a transition is running. This is the same pattern used in a couple of other
places.

(Arguably Cinematic should do this automatically but that's a bigger change.)