Simplify initial device detection
Previously, our code was reimplementing logic from InputHelper that maps joypad
names to a fixed set of types.

There is no need to do this: we can use the InputHelper.last_known_joypad_device
and InputHelper.last_known_joypad_index properties, which are set each time an
input event is processed.

The only difference in behaviour from this change is that if a joypad is
connected, the keyboard input hints will still be displayed by default until
input is processed from the joypad. In practice this won't happen: the player
will have used the joypad to choose (Re)start/Continue on the main menu, causing
these properties to be set! It only affects playing an individual scene during
development.