Extract Fray's End movement input hints to scene
Previously, the directional input hint controls were positioned
absolutely within Fray's End. This broke when we changed the game
resolution.

Move these to a new scene. Position the keyboard arrow icons using a
GridContainer: 2 rows, 3 columns, dummy controls in the top-left and
top-right spaces.

Stack the joypad dpad input hints (of which there are 4) in one place
using a CenterContainer. (This is a bit of a weird use of this container
but I couldn't find a better one!)

Position the two groups side-by-side using an HBoxContainer. (They'll
never both be visible so their relative position doesn't matter.)

Back in Fray's End, use another HBoxContainer to pack the
MovementInputHints next to the InteractInputHints (which I renamed for
consistency, and turned into a CenterContainer for reasons I don't
recall). The intermediate container is necessary because we use modulate
on the parent to fade it in and out of view, and modulate on the texture
for its pressed state.

Now anchor that HBoxContainer to the bottom left of the viewport.

Fixes https://github.com/endlessm/threadbare/issues/997