shjourney: Replace xylophone sound effects
In the main game, we have new, better-sounding sequence puzzle sounds.
And unlike the xylophone samples that were used before, we have a full
scale and a half, including all semitones.

Replace the xylophone sounds in the dragon puzzle with handbell sounds.
Change the scale from C major (C D E F G A) to C minor (C D D♯=E♭ F G
G♯=A♭): this matches the background music in this level, which is in C
minor.