REUSE.toml: Reindent with tabs
If you open and save this file in Godot's editor, it will replace the
spaces with tabs. It's really easy to do this accidentally because this
file is at the top-level so is shown by default in the File System
panel.

Godot does have an option to use spaces, but it's global to all file
types and can't be set in project.godot.

Reindent the file with tabs, adjusting editorconfig accordingly, so that
accidentally opening the file doesn't cause accidental edits to the
file.