Move Flatpak manifest to linux/ directory
Having it at the top level of the tree makes it very prominent when this
is a relatively niche file. It also makes it easy to accidentally edit
it in Godot and change its indentation.

Move it to the linux/ directory, which has a `.gdignore` file so is not
shown at all in the Godot file system view. Adjust the CI workflow
accordingly.