ci: Don't build branches/PRs that don't affect game
Previously, pushing any branch or opening any PR would cause the game to be
built.

There are many files in the repository which don't affect the game itself;
building the game for changes to these files is wasteful.

Configure the push and PR triggers to ignore branches/PRs that only affect such
files. I compiled the list of paths by hand.