PathWalkBehavior: Fix check for pointy in closed path
Curve2D has only a control point leading out for the first point, and
only an control point leading in for the last point. This special case
wasn't considered before. Add a helper function _is_curve_smooth() which
normally receives the in and out control points of the same vertex, but
for the special case receives the control points of the two endings.

With this, a path that is completely smooth (e.g. a circle path) should
not emit the pointy_path_reached signal at all.