PathWalkBehavior: Fix calculation bugs around the endings
To move the character along a path, we convert back and forth global
positions to curve offsets:

1. The current offset is determined. This is the distance in pixels from
the curve starting point to the character global position (actually the
closest point in the path to it).

2. A new offset is obtained by adding a delta (in pixels) depending on
the walk speed and direction.

3. A target position is determined from the new offset.

4. The character velocity is determined as a Vector2 that goes from the
current position to the target position, with walk speed as length.

5. Finally move_and_slide() is called.

6. We check if any "pointy" point was crossed, to emit the corresponding
signal.

The previous implementation had 2 issues:

First, the new offset calculation wasn't considering that it must be
clamped between zero and the path length, and that it should go in
circles if the path is closed. For going in circles, use fposmod().

Second, the new offset can't be used for testing if a "pointy" point
was crossed, because it represents the target, not the real, reached
position. Use the character global position after move_and_slide()
instead. And to simplify testing, calculate a test offset that considers
if the initial position (offset zero) was crossed.

This fixes a bug where a character was not reporting one point as
"pointy". And a bug where a character was stuck in a point flipping the
direction at every call to _physics_process(), unable to continue
walking.