shjourney: Remove sprint animation from fighting player
This quest has two different SpriteFrames for the player - one used in
the scene where he fights (template_playerFight.tres), and the other
used in the other scenes (tim_sprite_frames.tres).

tim_sprite_frames defines a "run" animation with different frames to the
"walk" animation. "run" is used by the player scene, if it is present.

template_playerFight defines a "sprint" animation, with the same frames
as the "walk" animation. "sprint" is not referenced anywhere else in the
project.

Remove the unused "sprint" animation.