shjourney: Remove TileSetAtlasSources with no texture assigned
The many tilesets in this StoryQuest have a large number of TileSetAtlasSources
which have no texture assigned, but do have tiles created. This causes the
following error to be shown, once per tile, whenever one of these tilesets is
loaded in the editor, or exported:

    ERROR: Cannot create tile. The tile is outside the texture or tiles are already present in the space the tile would cover.
    ERROR: The TileSetAtlasSource atlas has no tile at (3, 2).
    ERROR: TileSetAtlasSource has no tile at (3, 2).

Remove all such atlas sources from the tilesets. This doesn't resolve all of
these errors, but it does resolve some of them.

https://github.com/endlessm/threadbare/issues/1306