Guard: Fix flipping the sprite horizontally
In 57b0a2e5 I removed more than necessary:

```
if not is_zero_approx(velocity.x):
    animated_sprite_2d.flip_h = velocity.x < 0
```

Bring it back, but now using CharacterSpriteBehavior and
CharacterAnimationPlayerBehavior nodes.

When the guard is alerted, disable the node controlling the
AnimationPlayer for walking / being idle, so it can be disabled to play
the alerted animation.

Inside the scene file, the position of the idle animation subresource
changes for no reason.