Remove unused xylophone samples
These have now been replaced with handbells in every scene that is currently
merged, and so are unused.

However, we know that there are forks of the game with StoryQuests that use the
old samples, and these may in due course be submitted upstream. To help with
this, take the UIDs from the old xylophone samples and assign them to the
corresponding pitches of the new samples, updating the scenes that use them
accordingly. By doing this, if a new StoryQuest is submitted which uses the
xylophone samples, Godot will be able to use the UID to find the new sound.

Having done this, remove the old samples and their entry in REUSE.toml.