linux: Fix AppStream metainfo/releases licenses
Previously the comment at the top of the file claimed that these files
are under CC-BY-4.0. But AppStream actually has a field that describes
the license of the metadata, and in this we have:

    <metadata_license>CC0-1.0</metadata_license>

While both licenses are on the AppStream specification's recommended
list of metadata licenses, I think we can err on the side of the more
permissive license here.

https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-metadata_license