Prevent intro dialogue from replaying on respawn in stealth challenges
Previously, if a stealth challenge had introductory dialogue played by a
Cinematic node, if the player was caught the dialogue would be played
again each time the player respawned.

Add a flag in GameState that indicates whether the introductory dialogue
for the current scene has been played. Set it from Cinematic. Clear it
when the current scene is changed.

The flag is not persisted across save & load.

Resolves https://github.com/endlessm/threadbare/issues/949

Co-authored-by: Will Thompson <wjt@endlessaccess.org>