FillingBarrel: Rename fill() to increment()
And allow passing the amount to increment. Default to increment by one
as before. Also rename the corresponding animation.

The "El Juguete Perdido" story quest has the barrel scene duplicated
just for changing a collision shape, so do the animation rename there
too.

While testing the duplicated barrel, the editor changed the combat level
of that story quest, so commit that as well.