Move template tiles out of NO_EDIT StoryQuest
These tiles are not actually used by the NO_EDIT (née template)
StoryQuest. Including them in the NO_EDIT quest means extra work to
rename these files, for no real benefit.

Instead, move the image files to the assets directory, and the tileset
(plus its test scene) to a new `tiles/` top-level directory. Split the
tileset in 2 - one for the floor and one for the elevation - and update
the test scene accordingly.

Resolves https://github.com/endlessm/threadbare/issues/1335