storyquest_bootstrap: Add crude progress indicator
Previously, once you pressed the button in the dialog, the dialog
disappeared immediately; but duplicating the StoryQuest NO_EDIT template
can take some time.

In addition, while working on this plugin, I occasionally hit a state
where the reimport step got stuck. Previously there was no indication
that the bootstrap process was still, allegedly, ongoing.

Add an indeterminate progress indicator. Show it, and make the other
controls in the dialog noninteractive, when the button is pressed. Only
hide the dialog once the process is complete. Don't allow the dialog to be
closed while the copy is ongoing.