storyquest_bootstrap: Copy template recursively
Previously this plugin only copied the opening and closing cutscenes, and
relinked them. We would like the plugin to copy the whole NO_EDIT template,
including the 3 challenges.

The previous approach was to hard-code which elements of the two scenes need to
be copied and relinked. This works fine for these simple scenes but to handle
the complex interactive scenes we need a more general approach.

Define a copy() function which copies a resource, delegating to per-type special
cases to handle each type of resource that we need to handle: Quest,
PackedScene, CompressedTexture2D, DialogueResource, and SpriteFrames. For
scenes, check every property of every node in the scene which could be either a
Resource or string UID, and recursively call copy() if it points to a resource
in the NO_EDIT template.

Since this is a graph, not a tree, maintain a cache of source UID to copied
resource, and consult this before copying any given resource.

Kick off the process by copying quest.tres, which is the root of the graph.

Resolves https://github.com/endlessm/threadbare/issues/1214