Player: Set wall min slide angle to one degree
So the player slides nicely when colliding with non-walkable floor tiles
or any wall tile. The default of 15 degrees made the player capsule
collision shape get stuck in tile corners.

Zero can't be used, otherwise the player moves up when trying to go left
or right colliding with a vertical shape.