Make the barley bigger
Previously the barley asset in Fray's End was scaled to 50% of its natural size.

I actually think it looks better at full size. A barley plant should be the same
height as a large bush but previously it was much smaller. And with this larger
size, StoryWeaver can hide behind it!