Make the tomatoes bigger
Previously the tomato plants in Fray's End were scaled to 50%.

I think these look better full size. Tomato plants can grow pretty tall,
and having them at full size means we can appreciate this lovely
hand-drawn asset.

Adjust the position of one of the plants in Fray's End slightly so that
it is slightly offset from its neighbor. Similarly adjust the position
of the farmer to not overlap the tomato.