HookControl: Use a timer to allow D-pad diagonal aiming
The previous solution of just returning from _unhandled_input when a key
is released is not enough. Because when the player is walking in a
horizontal or vertical direction, it may continue aiming diagonally.

Factor out the axis calculation for rotating the aim control.

Add a Timer node and connect its timeout to the axis calculation.

Start / reset the timer when any key of the D-pad is released.

Fix https://github.com/endlessm/threadbare/issues/1366